Scripts to be installed to transform a Centipede GSNSS RTK base into a portable base
The GNSS RTK Centipede base modified with these scripts is not meant to be sending the RTCM correction data to the public caster, but rather to be used locally. Some applications require a very stable GNSS location (centimetric), but not an accurate absolute GNSS location. The portable GNSS RTK base can be deployed on the field easely, it automatically looks for proper satellite fix and then automatically sets the base coordinates to be used for RTCM RTK corrections. This portable GNSS RTK base enables very short range RTK corrections and very good location fix stability when the absolute location is not required.
https://docs.centipede.fr/